tests in new file, remove Client(None), rework multiple start() for multidb
